requirements.txt
Django==1.11.28
django-filter
Markdown==2.1.1
argparse==1.2.1
djangorestframework==3.7.7
psycopg2==2.8.5
six==1.15.0
wsgiref==0.1.2